fix rising edge of axis not working
setting the old axis equal to the current axis doesn't copy the array it instead references the same array so when we changed the values in the current Axis array it was also reflected when the array was accessed from the oldAxis field. To fix this I now explicitly copy the array.